Phase 3: Refactor Library Actions & Page
